fix tidy check & format fix

+++ src/include/duckdb/common/bit_utils.hpp
@@ -92,7 +92,7 @@ struct CountOnes<uint64_t> {
 		return idx_t(__popcnt(static_cast<unsigned int>(value)) + __popcnt(static_cast<unsigned int>(value >> 32)));
 #endif
 #else
-		return idx_t(__builtin_popcountll(static_cast<unsigned long long>(value)));
+		return idx_t(__builtin_popcountll(value));
 #endif
 	}
 };
